Address final review: empty-range result invariant, inBand overflow, doc terminology (CF-454)
- EventMutualInformationResult rejects an empty sample range carrying defined
  metrics or nonzero effective bins.
- WarpingWindow.inBand computes the index distance in long to avoid int
  overflow.
- Cross-module Javadoc link removed; current-or-future target-window
  terminology aligned across the example, README, and CHANGELOG.

+++ ta4j-core/src/main/java/org/ta4j/core/indicators/statistics/WarpingWindow.java
@@ -61,6 +61,6 @@ public record WarpingWindow(int radius, boolean unrestricted) {
      * @return {@code true} when the cell lies inside the alignment band
      */
     public boolean inBand(int firstIndex, int secondIndex) {
-        return unrestricted || Math.abs(firstIndex - secondIndex) <= radius;
+        return unrestricted || Math.abs((long) firstIndex - secondIndex) <= radius;
     }
 }
